Implement fixes

+++ IzvorniKod/back/portane/src/main/java/fer/portane/facade/impl/ArticleFacadeImpl.java
@@ -6,10 +6,12 @@ import fer.portane.form.ArticleForm;
 import fer.portane.mapper.ArticleArticleDtoMapper;
 import fer.portane.model.Article;
 import fer.portane.service.*;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Objects;
 
 @Component
 public class ArticleFacadeImpl implements ArticleFacade {
@@ -31,9 +33,16 @@ public class ArticleFacadeImpl implements ArticleFacade {
     private SeasonService seasonService;
     @Autowired
     private ClosetCustomComponentService closetCustomComponentService;
+    @Autowired
+    private AuthService authService;
+    @Autowired
+    private ClosetService closetService;
+
     @Override
+    @Transactional
     public ArticleDto save(ArticleForm form) {
-        Article article = new Article();
+        Article article = articleService.findById(form.getId());
+
         article.setLabel(form.getLabel());
         article.setPicture(form.getPicture());
         article.setPublic(form.isPublic());
@@ -48,14 +57,23 @@ public class ArticleFacadeImpl implements ArticleFacade {
         article.setPrimaryColor(colorService.findById(form.getPrimaryColorId()));
         article.setSecondaryColor(colorService.findById(form.getSecondaryColorId()));
 
-        article.setSeasons(form.getSeasonIds().stream().map(seasonService::findById).toList());
-        article.setStyles(form.getStyleIds().stream().map(styleService::findById).toList());
+        article.setSeason(seasonService.findById(form.getSeasonId()));
+        article.setStyles(form.getStyleIds().stream().map(styleService::findById).filter(Objects::nonNull).toList());
 
         return articleDtoMapper.toDto(articleService.save(article));
     }
 
     @Override
     public void deleteById(Long id) {
+        Article article = articleService.findById(id);
+
+        if (!Objects.equals(article.getClosetCustomComponent().getCloset().getId(), authService.getAuthenticatedUser().getId())){
+            throw new RuntimeException("You are not allowed to delete this article");
+        }
+
+        if (article == null) {
+            throw new RuntimeException("Article with id = " + id + " not found");
+        }
         articleService.deleteById(id);
     }
